Protecting Resources with Authorization

+++ backend/routes/posts.js
@@ -67,8 +67,12 @@ router.put('/:id', checkAuth, multer({ storage }).single("image"), (req, res, ne
     imagePath: imagePath
   });
   console.log(post);
-  Post.updateOne({ _id: req.params.id }, post).then(result => {
-    res.status(200).json({ message: 'Update successful!' })
+  Post.updateOne({ _id: req.params.id, creator: req.userData.userId }, post).then(result => {
+    if (result.nModified > 0) {
+      res.status(200).json({ message: 'Update successful!' })
+    } else {
+      res.status(401).json({ message: 'Not authorized!' })
+    }
   });
 });
 
@@ -107,9 +111,13 @@ router.get('/:id', (req, res, next) => {
 });
 
 router.delete('/:id', checkAuth, (req, res, next) => {
-  Post.deleteOne({ _id: req.params.id }).then(
+  Post.deleteOne({ _id: req.params.id, creator: req.userData.userId }).then(
     result => {
-      res.status(200).json({ message: 'Post deleted!' });
+      if (result.n > 0) {
+        res.status(200).json({ message: 'Deletion successful!' })
+      } else {
+        res.status(401).json({ message: 'Not authorized!' })
+      }
     }
   );
 });
